informe_claki_mail _d/requirements.txt
altair==4.1.0
appnope==0.1.2
argon2-cffi==20.1.0
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
backoff==1.11.1
beautifulsoup4==4.9.3
bleach==3.3.1
boltons==20.2.1
Brotli==1.0.9
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==7.1.2
click-spinner==0.1.10
colorlog==4.8.0
cssselect2==0.4.1
dacite==1.6.0
datapane==0.11.11
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
dominate==2.6.0
entrypoints==0.3
enum34==1.1.10
face==20.1.1
fonttools==4.26.2
furl==2.1.2
glom==20.11.0
greenlet==1.1.1
html5lib==1.1
idna==3.2
importlib-resources==5.2.2
ipykernel==6.2.0
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kaleido==0.2.1
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
micawber==0.5.4
mistune==0.8.4
monotonic==1.6
munch==2.5.0
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.3
numpy==1.21.2
orderedmultidict==1.0.1
packaging==20.9
pandas==1.3.2
pandocfilters==1.4.3
parso==0.8.2
pdfkit==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
plotly==5.2.1
posthog==1.4.2
prometheus-client==0.11.0
prompt-toolkit==3.0.20
ptyprocess==0.7.0
pyarrow==3.0.0
pycparser==2.20
pydantic==1.8.2
pydyf==0.1.0
Pygments==2.10.0
PyMySQL==1.0.2
pyparsing==2.4.7
pyphen==0.11.0
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.2.1
qrcode==7.2
qtconsole==5.1.1
QtPy==1.10.0
reportbro-fpdf==1.7.11
reportbro-lib==2.0.0
requests==2.26.0
requests-toolbelt==0.9.1
scipy==1.7.3
Send2Trash==1.8.0
simpleeval
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.22
stringcase==1.2.0
tabulate==0.8.9
tenacity==8.0.1
terminado==0.11.1
testpath==0.5.0
tinycss2==1.1.0
toolz==0.11.1
tornado==6.1
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.6
validators==0.18.2
wcwidth==0.2.5
weasyprint==53.0
webencodings==0.5.1
widgetsnbextension==3.5.1
wkhtmltopdf==0.2
XlsxWriter==3.0.1
zipp==3.5.0
zopfli==0.1.8
